refactor(search): extract time filter helpers into _timeFilter.ts
